apps/ai/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# AI & LLM
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.20

# Document Processing
pdfplumber>=0.10.0
docx2txt>=0.8
beautifulsoup4>=4.12.0
requests>=2.31.0

# NLP & Embeddings
sentence-transformers>=2.2.0

# Audio Processing
faster-whisper>=0.9.0

# Database & Storage
supabase>=2.0.0
boto3>=1.34.0

# Search & Communication
googlesearch-python>=1.2.0
resend>=0.6.0

# Environment & Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.0.0

# Code Analysis & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Monitoring & Logging
prometheus-client>=0.19.0

# Development Tools
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4